chore: create FUNDING.yml
I recently released [`cargo-fund`](https://users.rust-lang.org/t/cargo-fund-discover-funding-links-for-your-projects-dependencies/43251?u=acfoltzer), and noticed that while there is sponsorship information in this repo, but it's not yet encoded in a machine-readable way. Adding this metadata will add a "Sponsor" button to the repo, and allow `cargo-fund` to show the project's sponsorship link. For more about this file format, see the [Github docs](https://help.github.com/en/github/administering-a-repository/displaying-a-sponsor-button-in-your-repository).

Updated from the patch in #1944 with the Open Collective funding link.