fix: Remove overriden occurrences as we go
This allows two overriding args to interact with each other mid-line.
This was broken in c19cc01 / #1154.  Before that, we duplicated the override
logic in several places.

Now we plug into the start of a new arg which allows us to do this
incrementally without making the logic complex or inefficient, thanks to
prior refactors.

Fixes #3217